Revert "IRT-2892 implemented clients for REST APi and script collections. Ref…"
