+++ build.gradle
@@ -8,7 +8,7 @@ configurations.all {
 }
 
 group = 'com.github.figaf'
-version = '1.18'
+version = '1.17'
 
 repositories {
     mavenLocal()
