+++ src/main/java/com/figaf/integration/cpi/entity/designtime_artifacts/CreateOrUpdateCpiArtifactRequest.java
@@ -1,6 +1,5 @@
 package com.figaf.integration.cpi.entity.designtime_artifacts;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -14,29 +13,20 @@ import java.util.List;
 @Getter
 @Setter
 @ToString
-@AllArgsConstructor
 public abstract class CreateOrUpdateCpiArtifactRequest {
 
     private String id;
     private String name;
     private String description;
-    private AdditionalAttributes additionalAttrs;
+    private String type;
+    private AdditionalAttributes additionalAttrs = new AdditionalAttributes();
     private String fileName;
-
-    public abstract String getType();
-
-    public AdditionalAttributes getAdditionalAttrs() {
-        if (additionalAttrs == null) {
-            additionalAttrs = new AdditionalAttributes();
-        }
-        return additionalAttrs;
-    }
 
     @Getter
     @ToString
     public static class AdditionalAttributes {
 
-        private final List<String> source = new ArrayList<>();
-        private final List<String> target = new ArrayList<>();
+        private List<String> source = new ArrayList<>();
+        private List<String> target = new ArrayList<>();
     }
 }
