+++ src/main/java/com/figaf/integration/cpi/entity/designtime_artifacts/CreateOrUpdateIFlowRequest.java
@@ -1,30 +1,20 @@
 package com.figaf.integration.cpi.entity.designtime_artifacts;
 
-import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
 import lombok.ToString;
-
-import static com.figaf.integration.cpi.entity.designtime_artifacts.CpiArtifactType.IFLOW;
 
 /**
  * @author Nesterov Ilya
  */
+@Getter
+@Setter
 @ToString
 public class CreateOrUpdateIFlowRequest extends CreateOrUpdateCpiArtifactRequest {
 
-    @Builder
-    public CreateOrUpdateIFlowRequest(
-        String id,
-        String name,
-        String description,
-        AdditionalAttributes additionalAttrs,
-        String fileName
-    ) {
-        super(id, name, description, additionalAttrs, fileName);
-    }
-
     @Override
     public String getType() {
-        return IFLOW.getQueryTitle();
+        return "IFlow";
     }
 
 }
